move post body validation to service
